rename: 공개 여부 변경
